SB-34 (Refactor) : BaseEntity Package 변경
SB-34 (Refactor) : BaseEntity Package 변경

BaseEntity Package 변경
- global -> db
