+++ db/src/main/java/db/domain/goods/GoodsRepository.java
@@ -1,4 +1,4 @@
-package db.domain.goods.enums;
+package db.domain.goods;
 
 import db.domain.goods.GoodsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
